Fix audio messages parsing in vk

+++ BridgeBotNext/Attachments/AnimationAttachment.cs
@@ -9,7 +9,7 @@ namespace BridgeBotNext.Attachments
             string fileName = null,
             long fileSize = 0,
             string mimeType = null,
-            int duration = 0,
+            ulong duration = 0,
             int width = 0,
             int height = 0
         ) : base(url, meta, caption, fileName, fileSize, mimeType ?? "video/mp4", null, duration, width, height)
