+++ BridgeBotNext/Attachments/VoiceAttachment.cs
@@ -4,7 +4,7 @@ namespace BridgeBotNext.Attachments
     {
         public VoiceAttachment(string url = null, object meta = null, string caption = null, string fileName = null,
             long fileSize = 0,
-            string mimeType = null, string title = null, int duration = 0) : base(url, meta, caption, fileName,
+            string mimeType = null, string title = null, ulong duration = 0) : base(url, meta, caption, fileName,
             fileSize,
             mimeType, "audio/ogg", title, null, duration)
         {
